FE/Qt: A bunch of casting build-warnings.
svn:sync-xref-src-repo-rev: r100762

+++ src/VBox/Frontends/VirtualBox/src/VBoxFBOverlay.cpp
@@ -1,4 +1,4 @@
-/* $Id: VBoxFBOverlay.cpp 56126 2015-05-28 10:02:59Z ramshankar.venkataraman@oracle.com $ */
+/* $Id: VBoxFBOverlay.cpp 56217 2015-06-03 11:50:25Z sergey.dubov@oracle.com $ */
 /** @file
  * VBox Qt GUI - VBoxFBOverlay implementation.
  */
@@ -4431,10 +4431,10 @@ bool VBoxQGLOverlay::onNotifyUpdate(ULONG uX, ULONG uY,
         const double yScaleFactor = mSizeInfo.visualState() == UIVisualStateType_Scale ?
                                     (double)scaledSize.height() / mSizeInfo.height() : dScaleFactor;
         /* Adjust corresponding viewport part: */
-        rect.moveTo(floor((double)rect.x() * xScaleFactor) - 1,
-                    floor((double)rect.y() * yScaleFactor) - 1);
-        rect.setSize(QSize(ceil((double)rect.width()  * xScaleFactor) + 2,
-                           ceil((double)rect.height() * yScaleFactor) + 2));
+        rect.moveTo((int)floor((double)rect.x() * xScaleFactor) - 1,
+                    (int)floor((double)rect.y() * yScaleFactor) - 1);
+        rect.setSize(QSize((int)ceil((double)rect.width()  * xScaleFactor) + 2,
+                           (int)ceil((double)rect.height() * yScaleFactor) + 2));
     }
 
 #ifdef Q_WS_MAC
@@ -4444,10 +4444,10 @@ bool VBoxQGLOverlay::onNotifyUpdate(ULONG uX, ULONG uY,
         const double dBackingScaleFactor = darwinBackingScaleFactor(mpViewport->window());
         if (dBackingScaleFactor > 1.0)
         {
-            rect.moveTo(floor((double)rect.x() / dBackingScaleFactor) - 1,
-                        floor((double)rect.y() / dBackingScaleFactor) - 1);
-            rect.setSize(QSize(ceil((double)rect.width()  / dBackingScaleFactor) + 2,
-                               ceil((double)rect.height() / dBackingScaleFactor) + 2));
+            rect.moveTo((int)floor((double)rect.x() / dBackingScaleFactor) - 1,
+                        (int)floor((double)rect.y() / dBackingScaleFactor) - 1);
+            rect.setSize(QSize((int)ceil((double)rect.width()  / dBackingScaleFactor) + 2,
+                               (int)ceil((double)rect.height() / dBackingScaleFactor) + 2));
         }
     }
 #endif /* Q_WS_MAC */
